adjust root directory structure
